refactor : Admin 관련 Repository 이름 수정
