+++ src/main/java/com/flab/shoeauction/domain/users/user/UserRepository.java
@@ -1,9 +1,10 @@
 package com.flab.shoeauction.domain.users.user;
 
+import com.flab.shoeauction.domain.users.admin.AdminRepository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryByAdmin {
+public interface UserRepository extends JpaRepository<User, Long>, AdminRepository {
 
     boolean existsByEmail(String email);
