--- src/main/java/com/flab/shoeauction/domain/users/user/UserRepositoryByAdmin.java
@@ -1,11 +1,0 @@
-package com.flab.shoeauction.domain.users.user;
-
-import com.flab.shoeauction.controller.dto.UserDto.UserSearchCondition;
-import com.flab.shoeauction.controller.dto.UserDto.UserListResponse;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface UserRepositoryByAdmin {
-
-    Page<UserListResponse> searchByUsers(UserSearchCondition searchRequest, Pageable pageable);
-}
