+++ src/main/resources/templates/gallery.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
+<head>
+  <meta charset="UTF-8">
+  <title>S3 파일 업로드 테스트</title>
+</head>
+<body>
+<h1>파일 업로드</h1> <hr>
+
+<form th:action="@{/gallery}" method="post" enctype="multipart/form-data">
+  제목 : <input type="text" name="title"> <br>
+  파일 : <input type="file" name="file"> <br>
+  <button>등록하기</button>
+</form>
+
+</body>
+</html>
